Start over to use Emacs customization.
I have decided to use Emacs customization (M-x customize) as much as
possible, to reduce the amount of code that must be dealt with manually.

If one has to believe a comment is this thread:

    https://www.reddit.com/r/emacs/comments/1udtd1/starting_emacs_with_preludestarter_kits_scares_me/

doing so is also more efficient than applying such customizations in
init.el. But I have not made any test and I have not enough experience
to confirm or deny that. Anyway, using Emacs Customization simplifies my
life as an Emacs (re)starter *a lot*, so I'm going to use it.