Automaticallly install missing packages.
Implement functions to check whether a set of "wanted" packages are
already installed. If some package is missing, download it from a remote
repository using `package-install`. In downloading fails, emit a log
message, but do not prevent downloading further packages.

I tried to perform package checks in a function called from
after-init-hook, but for some reasons I was getting:

    package.el is not yet initialized

I wanted to avoid calling `package-initialize` in my init file, but that
seems to be the simplest thing to do. I will unset
`package-enable-at-startup` in a later commit.

Also, I had to set `package-archives` in `lifepillar-packages.el` rather
than through Customize, because settings from Customize are loaded at a
later moment. I cannot (well, I do not want to) load custom.el earlier,
because I would not be able to have a theme set through Customize loaded
at startup (themes set in `custom-enabled-themes` are loaded with
`load-theme`, which does not find the theme's files if packages have not
been initialized). All in all, I think that this part of Emacs is quite
a mess.

The only packages I have added for now are a few themes. Zenburn is the
default theme (I have tried the most popular Solarized theme from MELPA,
but it does not seem to work in the terminal).